+++ repos/messenger/src/parent/playwright/index.js
@@ -1,1 +1,2 @@
 export * from './elementHandle'
+export * from './page'
