+++ repos/messenger/src/parent/utils/dom/index.js
@@ -1,0 +1,1 @@
+export * from './withSelector'
