+++ repos/messenger/src/parent/utils/index.js
@@ -1,0 +1,1 @@
+export * from './dom'
